% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClinicalTrial.AE}
\alias{ClinicalTrial.AE}
\title{Clinical Trial: Adverse events}
\format{
A data frame with 11 rows and 6 variables:
\describe{
  \item{id}{Patient id}
  \item{time}{Time of an adverse event (AE)}
  \item{event}{Type of adverse event (AE)}
  \item{Sex}{Patient Sex}
  \item{Age}{Age of patient at trial entry date}
  \item{Related}{Likelihood the treatment is related to the adverse event}

}
}
\usage{
ClinicalTrial.AE
}
\description{
A dataset containing the adverse event information from a simulated clinical trial
}
\keyword{datasets}
