% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClinicalTrial.Arm}
\alias{ClinicalTrial.Arm}
\title{Clinical Trial: Treatment}
\format{
A data frame with 53 rows and 6 variables:
\describe{
\item{id}{Patient id}
  \item{Arm}{Treatment Arm}
  \item{End_trt}{Time since enrollment to the end of treatment, in months}
  \item{Continued_treatment}{Continued treatment past end of follow up}
  \item{Sex}{Patient Sex}
  \item{Age}{Age of patient at trial entry date}

}
}
\usage{
ClinicalTrial.Arm
}
\description{
A dataset containing the treatment arm information from a simulated clinical trial
}
\keyword{datasets}
