% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClinicalTrial.Response}
\alias{ClinicalTrial.Response}
\title{Clinical Trial: Response}
\format{
A data frame with 36 rows and 7 variables:
\describe{
  \item{id}{Patient id}
  \item{Response_start}{Time of starting response, in months since enrollment}
  \item{Response_end}{Time of ending response, in months since enrollment}
  \item{Response}{Type of response, CR = Complete response, and PR = Partial response}
  \item{Continued_response}{Continued response past end of follow up}
  \item{Sex}{Patient Sex}
  \item{Age}{Age of patient at trial entry date}

}
}
\usage{
ClinicalTrial.Response
}
\description{
A dataset containing the response information from a simulated clinical trial
}
\keyword{datasets}
