% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ClinicalTrial.Stage}
\alias{ClinicalTrial.Stage}
\title{Clinical Trial: Stage}
\format{
A data frame with 36 rows and 2 variables:
\describe{
  \item{id}{Patient id}
  \item{Stage}{Patients clinical stage at enrollment of the study (either Early Stage or Late Stage)}

}
}
\usage{
ClinicalTrial.Stage
}
\description{
A dataset containing the Stage information from a simulated clinical trial
}
\keyword{datasets}
