% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimplot.R
\name{line_df_to_point_df}
\alias{line_df_to_point_df}
\title{Formats a dataframe of line to add points}
\usage{
line_df_to_point_df(df_lines, start = "start", end = "end", cont = NULL)
}
\arguments{
\item{df_lines}{a dataframe}

\item{start}{start column name}

\item{end}{end column name}

\item{cont}{continue column name}
}
\value{
a dataframe in a format for adding points to a swimmers plot
}
\description{
This function formats a dataframe; used with \code{\link{swimmer_lines}}
}
