% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swimplot.R
\name{swim_plot_pattern}
\alias{swim_plot_pattern}
\title{Creating the base of a swimmers plot}
\usage{
swim_plot_pattern(
  df,
  id = "id",
  end = "end",
  name_fill = NULL,
  name_col = NULL,
  name_pattern = NULL,
  name_pattern_fill = NULL,
  id_order = "increasing",
  stratify = FALSE,
  base_size = 11,
  ...
)
}
\arguments{
\item{df}{a data frame}

\item{id}{column name for id, default is 'id'}

\item{end}{column name with the bar sizes (or bar end positions if bars change colour)}

\item{name_fill}{a column name to map the bar fill}

\item{name_col}{a column name to map the bar colour}

\item{name_pattern}{a column name to map the pattern}

\item{name_pattern_fill}{a column name to map the pattern fill}

\item{id_order}{order of the bars by id, default is "increasing", can also input
"decreasing", a column name, or the ids in an order.}

\item{stratify}{a list of column names to stratify by}

\item{base_size}{the base size for the plot, default is 11}

\item{...}{additional geom_col() arguments}
}
\value{
a swimmer plot with bars
}
\description{
This function allows you to create swimmers plots with bars, includes options to
have the bars change colours and create stratified plots
}
\seealso{
\code{\link{swimmer_points}} \code{\link{swimmer_lines}}  \code{\link{swimmer_lines}}  \code{\link{swimmer_points_from_lines}} \code{\link{swimmer_arrows}} \code{\link{swimmer_text}}
}
